Add pagination to filter dimensions
Uses MongoDB facets to return the nested array as well as a total count.

In order to determine if the original match was a success (i.e. if the
filter existed in the first place), an additional flag is added to the
facets. If the flag is set, then at least one result was returned by the
match.

Resolves: 5559