Ensure dimension name is unescaped before querying
Updated to decode the dimension name, since currently dimensions are
stored using their pretty name, e.g. `Number of Siblings`, and passed
in the URL as encoded (e.g.  `Number+of+Siblings`).

Until this is changed we need to unescape the dimension before
querying, or it doens't work if the name contains any special
characters/spaces.

Resolves: 5570 (https://trello.com/c/8xCrDyR1)